% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\docType{data}
\name{fittedModelRMR}
\alias{fittedModelRMR}
\title{Fitted RMR model.}
\format{
A list object obcontaining the fixed effects estimates derived from
  the brms model fit of RMR by mass, temperature and species. The entire
  model object is not provided as it is tens of Mb in size and is not
  strictly required.
}
\usage{
data(fittedModelRMR)
}
\description{
Fitted RMR model.
}
\author{
Andrew Jackson
}
\keyword{datasets}
