% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalist.R
\docType{data}
\name{Kmass}
\alias{Kmass}
\title{Cooling coefficients by body mass and species.}
\format{
An object of class `"data.frame"` containing three variables. 
The first column is a character vector of species names labelled "species". 
The second is labelled "M" is a numeric vector of body mass values in
log10(kg). The third is labelled "K" represents cooling coefficients.
}
\usage{
data(Kmass)
}
\description{
Taken from Nakamuru XYZH REF!
}
\author{
Andrew Jackson
}
\keyword{datasets}
