% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMRfromTa.R
\name{RMRfromTa}
\alias{RMRfromTa}
\title{Esimate Routine Metabolic Rate (RMR) from ambient water temperature (Ta)}
\usage{
RMRfromTa(m, Ta, meso, pars, kpars = NULL)
}
\arguments{
\item{m}{mass of fish in kg}

\item{Ta}{temperature of the ambient water}

\item{meso}{a logical value where TRUE indicates mesotherm and FALSE
indicates ectotherm}

\item{pars}{pars a vector of length 4 containing the coefficients of the
corresponding fitted linear regression}

\item{kpars}{a vector contain "a" the prefactor of the exponential scaling
equation and "b" the scaling exponent. Supplying NULL, the default, uses
default parameter. See `?kfun` for details.}
}
\description{
Esimate Routine Metabolic Rate (RMR) from ambient water temperature (Ta)
}
