% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T0fun.R
\name{T0fun}
\alias{T0fun}
\title{Estimate metabolic heat production T0 using a fitted linear model}
\usage{
T0fun(m, Tm, meso, pars)
}
\arguments{
\item{m}{body mass of the fish in kg}

\item{Tm}{body temperature of the fish}

\item{meso}{a switch parameter that is either TRUE for mesotherm (or regional
endotherm) or FALSE for ectotherm}

\item{pars}{the coefficients of the fitted regression model}

\item{omega}{the coefficient of mass to convert metabolic rate into temperature. Defaults to 10^0.5945}
}
\value{
a vector the same length as input m containing the corresponding estimated metabolic heat production parameter T0
}
\description{
Estimate metabolic heat production T0 using a fitted linear model
}
