% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyTempLambert.R
\name{bodyTempLambert}
\alias{bodyTempLambert}
\title{Solve the model for body temperature as a function of ambient water temp}
\usage{
bodyTempLambert(x, m, meso, pars)
}
\arguments{
\item{x}{ambient water temperature in C}

\item{m}{body mass in g}

\item{meso}{a switch parameter that is either TRUE for mesotherm (or regional
endotherm) or FALSE for ectotherm}

\item{pars}{the coefficients of the fitted regression model}
}
\value{
body temperature in C
}
\description{
Solve the model for body temperature as a function of ambient water temp
}
