% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRegressionK.R
\name{fitRegressionK}
\alias{fitRegressionK}
\title{Fit a log-log regression model to estimate K from body mass}
\usage{
fitRegressionK(dd = utils::data("Kmass"))
}
\arguments{
\item{dd}{a data.frame or tibble object comprising three columns: species, M
(log mass in ) and cooling coefficient K. Defaults to the included dataset
Kmass.}
}
\value{
a model object of a log-log regression
}
\description{
Fits a random intercept model of cooling coefficient by body mass with 
species as random factor
}
