% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfun.R
\name{kfun}
\alias{kfun}
\title{Estimate k from Nakamura's exponential scaling}
\usage{
kfun(m, kpars = c(a = 0.00173, b = -0.633))
}
\arguments{
\item{m}{body mass of the fish in g}

\item{kpars}{a vector contain "a" the prefactor of the exponential scaling
equation and "b" the scaling exponent}
}
\value{
an estimate of the cooling coefficient k
}
\description{
Estimate k from Nakamura's exponential scaling
}
\examples{
kfun(1000)
kfun(m = c(1, 10, 100, 1000), kpars = c(a = 0.00201, b = -0.617))
}
