% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfun.R
\name{kfun}
\alias{kfun}
\title{Estimate k from Nakamura's exponential scaling}
\usage{
kfun(m, a = 0.00161, b = -0.62)
}
\arguments{
\item{m}{body mass of the fish in g}

\item{a}{the prefactor of the exponential scaling equation}

\item{b}{the scaling exponent}
}
\value{
an estimate of the cooling coefficient k
}
\description{
Estimate k from Nakamura's exponential scaling
}
\examples{
kfun(1000)
kfun(m = c(1, 10, 100, 1000), a = 0.00161, b = -0.62)
}
