% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxAmbientTemp.R
\name{maxAmbientTemp}
\alias{maxAmbientTemp}
\title{Calculate the maximum value of water temperature that yields real numbers}
\usage{
maxAmbientTemp(m, meso, pars)
}
\arguments{
\item{m}{body mass of the fish in g}

\item{meso}{a switch parameter that is either TRUE for mesotherm (or regional
endotherm) or FALSE for ectotherm}

\item{pars}{the coefficients of the fitted regression model}
}
\value{
a vector of length m with corresponding max water temperatures in C
}
\description{
Calculate the maximum value of water temperature that yields real numbers
}
