% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictFishRMR-package.R
\docType{package}
\name{predictFishRMR-package}
\alias{predictFishRMR}
\alias{predictFishRMR-package}
\title{predictFishRMR: Estimates the resting metabolic rates of fish based on allometric scaling equations and water temperatures}
\description{
Observed relationships between body mass and cooling coefficients, coupled with a seperate relationship between body mass and temperature elevations above ambient water temperature are used to provide predictions of resting metabolic rates of fish using only body mass and whether they are mesotherms (or regional endotherms) or ectotherms. Additional functions are provided to estimate predicted maximum ambient water temperature that permits thermal equilibrium.
}
\keyword{internal}
