% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMRfromTa.R
\name{RMRfromTa}
\alias{RMRfromTa}
\title{Esimate Routine Metabolic Rate (RMR) from ambient water temperature (Ta)}
\usage{
RMRfromTa(m, Ta, meso, pars, kpars = NULL)
}
\arguments{
\item{m}{mass of fish in kg}

\item{Ta}{temperature of the ambient water}

\item{meso}{a logical value where TRUE indicates mesotherm and FALSE
indicates ectotherm}

\item{pars}{pars a vector of length 4 containing the coefficients of the
corresponding fitted linear regression}

\item{kpars}{a vector contain "a" the prefactor of the exponential scaling
equation and "b" the scaling exponent. Supplying NULL, the default, uses
default parameter. See `?kfun` for details.}
}
\value{
a vector the same length as `m` with corresponding RMR values
}
\description{
Esimate Routine Metabolic Rate (RMR) from ambient water temperature (Ta)
}
\examples{
RMRfromTa(m = c(100, 1000), Ta = 15, meso = T, pars = c(2.95, 0.83, 0.08, 1.23), kpars = c(a = 0.00201, b = -0.617))
}
